Refactor ServiceMetadata to ensure the row keeps its format
