+++ ts/components/services/ServiceMetadata/__tests__/InformationRow.test.tsx
@@ -1,0 +1,20 @@
+import React from "react";
+import { render } from "@testing-library/react-native";
+import { constNull } from "fp-ts/lib/function";
+
+import InformationRow from "../InformationRow";
+
+describe("the InformationRow component", () => {
+  it("should match the snapshot", () => {
+    expect(
+      render(
+        <InformationRow
+          value={"via Roma"}
+          label={"serviceDetail.fiscalCode"}
+          onPress={constNull}
+          hint={"clipboard.copyText"}
+        />
+      ).toJSON()
+    ).toMatchSnapshot();
+  });
+});
